+++ src/main/java/roomescape/payment/client/config/TossClientConfig.java
@@ -1,4 +1,4 @@
-package roomescape.payment.client;
+package roomescape.payment.client.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
