+++ src/test/java/roomescape/payment/client/TossPaymentGatewayTest.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
+import roomescape.payment.client.gateway.TossPaymentGateway;
 import roomescape.payment.exception.TossPaymentException;
 import roomescape.payment.service.dto.PaymentConfirmation;
 import roomescape.payment.service.dto.PaymentStatus;
